Lay out project base

+++ src/game/clue.rs
@@ -1,4 +1,6 @@
+/// Represents the action of giving a clue in a game
+#[derive(Clone, Debug, Eq, Hash, PartialEq)]
 pub struct Clue {
-    pub clue_type: usize,
-    pub clue_value: usize,
+    pub clue_type: u8,
+    pub clue_value: u8,
 }
